requirements.txt
numpy
pandas
pillow
scikit-image
scikit-learn
opencv-python
python-dotenv
fastapi
uvicorn
python-multipart
requests
huggingface-hub
